[services] Allow browsers to access prod identity from comm.software
Summary:
This diff updates the CORS config for the identity service in production to allow incoming connections from browsers rendering comm.software and www.comm.software.

Context for why we need this is in [ENG-9862](https://linear.app/comm/issue/ENG-9862/investigate-performance-issues-when-threadstore-changes).

Test Plan: Not sure how to test this besides deploying to prod...

Reviewers: varun, will, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13881