[shared] bump opaque-ke to 2.1.0-pre.1
Summary: This version of opaque-ke uses the latest version of voprf. The previous version of opaque-ke was using an older version of voprf that no longer works with the latest version of Rust. Hence, in order to upgrade to the latest version of Rust, we need to first upgrade opaque-ke.

Next diff will bump our Rust version from 1.76 to 1.82 (latest stable version)

Test Plan:

1. deployed custom identity image to staging with these changes
2. checked out master branch
3. successfully logged in to existing account on staging identity from native, web, keyserver
4. successfully registered and logged in to staging from native
5. switched back to feature branch
6. successfully logged in to existing account on staging identity from native, web, keyserver
7. successfully registered and logged in to staging from native

Reviewers: ashoat, will, bartek

Subscribers: