Upgrade to Rust 1.82
Summary:
Rust 1.81 introduced new [lifetime rules](https://github.com/rust-lang/rust/pull/117967), which opaque-ke, a crate we depend on, violated. The maintainers of that crate have introduced a new version that satisifies these new rules, and we have upgraded to this version in D13895. Now that we've upgraded the dependency, we should bump Rust to the latest stable version.

Depends on D13895

Test Plan: All crates build locally and in Docker (CI)

Reviewers: will, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13896