[lib] Extract getLastUpdatedTime to a separate file
Summary:
This will make a following diff a little easier.

Note that this isn't a pure move; I've removed the `isEmptyMediaMessage` check, which will be reimplemented inside a message spec in the next diff.

Depends on D13911

Test Plan: Flow. Also tested in combination with the rest of the stack

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13912