[lib] Introduce MessageSpec.getLastUpdatedTime
Summary:
This new method on the `MessageSpec` will allow us to address [ENG-9558](https://linear.app/comm/issue/ENG-9558/prevent-some-robotext-from-bumping-threads-in-the-chatlistdata-order). If it returns `null` or `undefined` that means that the message should not bump the thread's timestamp. If it returns a `Promise` that means we don't yet know if it should bump the thread's timestamp.

For now, in this diff we're ignoring any messages whose `MessageSpec.getLastUpdatedTime` returns a `Promise`, and assuming it's the same as if it returned `null`. In later diffs we'll update the logic to be smarter.

Depends on D13912

Test Plan: Tested in combination with the rest of the stack. See video in the last diff

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D13913