[lib][native][web] Update useSidebarInfos to include lastUpdatedAtLeastTime
Summary:
Before this diff, `useSidebarInfos` would return `lastUpdatedAtLeastTime` as `lastUpdatedTime`, as a temporary hack to help split up my work into smaller diffs.

In an earlier diff we updated `lastUpdatedTime` to be a `Promise`, and this allows us to avoid refactoring everything to handle the `Promise`.

In this diff, we address this temporary hack for sidebars, but for not for the `ChatThreadItem` itself (which will be handled later).

`useSidebarInfos` is used in places outside of `chat-selectors.js`, so this required some additional refactoring.

Depends on D13914

Test Plan: Tested in combination with the rest of the stack. See video in the last diff

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D13915