[lib] Reactions only bump thread timestamps when reacting to viewer
Summary: This addresses [ENG-9558](https://linear.app/comm/issue/ENG-9558/prevent-some-robotext-from-bumping-threads-in-the-chatlistdata-order) for reaction robotext messages. The implementation is broadly similar to D13828.

Depends on D13916

Test Plan:
I logged in as two users on web, and with the second user, tested reacting to the first user's messages and other messages. Confirmed that the timestamp was bumped for the first user only when reacting to their messages. I also tested this with a `sleep` to confirm that the timestamps would only be bumped after the promise resolved

[video]

Reviewers: tomek

Subscribers: