[native] make the list of NUX tips to be displayed configurable
Summary:
i'll be introducing a new tip that i want to display from the NUX tips overlay, separate from the tips that display on fist launch.

in order to do this, i need to be able to navigate to my new tip directly from the backdrop, as opposed to going through the 4 existing tips first.

in this diff, i remove nextTip from NUXTipParams. instead, in the NUX handler, we initialize an ordered list of tips and use this list to determine the next tip to display

https://linear.app/comm/issue/ENG-9913/make-the-list-of-nux-tips-to-be-displayed-configurable

Test Plan: on a fresh install, confirmed that the same 4 tips appeared as before, and that navigating through them was unchanged

Reviewers: ashoat, tomek

Reviewed By: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13917