[native_rust_library] update creating backup to return `backupID`
Summary:
[ENG-9656](https://linear.app/comm/issue/ENG-9656/native-changes-for-minimal-version-of-backup).

We need to store `backupID` in the store, it's needed in e.g. [ENG-9616](https://linear.app/comm/issue/ENG-9616/clients-store-informations-if-using-backup-or-v1-login#comment-f937dae3).

Depends on D13936

Test Plan:
This is the final diff for this stack so testing this end-to-end, mostly uploading User Keys, but also getting the latest Backup Info and Uploading full backup to test if this still works.
I was testing using physical iOS and Android, and both password and wallet users.
I was confirming everything with investigating DDB content.

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13937