[backup-client] avoid adding `attachments` when uploading User Keys
Summary:
[ENG-9656](https://linear.app/comm/issue/ENG-9656/native-changes-for-minimal-version-of-backup).

On backup service, we parse multipart in sequence, so we're going to read `attachments` and as a result set `siwe_backup_msg` to None even if this is present.

The alternative is to completely refactor how we parse Multipart on the service side - but this is quick and clean, as `attachments` are present only if we have User Data.

Test Plan: Run commtest

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13928