[native][lib] merge `BackupDataKeyMessage` with `DeviceListUpdateSuccess`
Summary:
[ENG-9905](https://linear.app/comm/issue/ENG-9905/update-secondary-device-login-to-match-whitepaper).

This diff merges types and moves retrieving keys to a different step.

This change is not backward compatible but this Flow hasn't been launched yet so we not breaking anything.

NOTE:
It's not that easy to slice this into diffs, so it's possible that in the middle of the stack protocol is not working, it needs to be reviewed as a whole.

Depends on D13964

Test Plan: Tested later in the stack - D13968

Reviewers: bartek, tomek, varun

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13965