[native][lib] remove `requestBackupKeys` from `SECONDARY_DEVICE_REGISTRATION_SUCCESS`
Summary:
[ENG-9905](https://linear.app/comm/issue/ENG-9905/update-secondary-device-login-to-match-whitepaper).

This makes no sense anymore, secondary will send this after registration, so we need to decide earlier based on the device type.

This change is not backward compatible but this flow hasn't been launched yet so we not breaking anything.

NOTE:
It's not that easy to slice this into diffs, so it's possible that in the middle of the stack protocol is not working, it needs to be reviewed as a whole.

Depends on D13965

Test Plan: Tested later in the stack - D13968

Reviewers: bartek, tomek, varun

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13966