[lib] temporary remove code performing backup restore
Summary:
[ENG-9905](https://linear.app/comm/issue/ENG-9905/update-secondary-device-login-to-match-whitepaper).

This is the first step of merging `BACKUP_DATA_KEY_MESSAGE` and `DEVICE_LIST_UPDATE_SUCCESS`. Initially, I wanted to make this code part of `DEVICE_LIST_UPDATE_SUCCESS`, after `performLogIn` is done - that is how this will work in the future. However, we're now running on a limited version of backup and there isn't anything to restore. Going to mention this in [ENG-992](https://linear.app/comm/issue/ENG-9921/revert-d13933-capturing-logs) in order to revert this.

NOTE:
It's not that easy to slice this into diffs, so it's possible that in the middle of the stack protocol is not working, it needs to be reviewed as a whole.

Depends on D13963

Test Plan: There is no restore attempt even if backup keys are sent to secondary device.

Reviewers: bartek, tomek, varun

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13964