[native] avoid showing SIWE backup message screen for non-primary devices
Summary:
[ENG-9905](https://linear.app/comm/issue/ENG-9905/update-secondary-device-login-to-match-whitepaper).

This a fix for [ENG-9917](https://linear.app/comm/issue/ENG-9917/siwe-user-is-navigated-to-createmissingsiwebackupmessage-after).

Depends on D13967

Test Plan:
This is the final diff so testing this end-to-end.
I was scanning the QR code using my physical iOS device, secondary was the web and other iOS physical devices. I was testing this on both passwords and wallet users.
Things I checked:
- Secondary logged in
- Secondary received backup data (but ignored this for now)
- On primary, the camera was dismissed and Secondary was added to the list
- Tunnel broker connection (auth) created
- Device list updates are sent

Things I didn't test (not sure if I should):
- keyserver auth
- scanning on Android (I don't have access to this device when creating this stack, I can do it later in the week)

Reviewers: bartek, tomek, varun

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13968