[keyserver] implement invite link button to frog hono route
Summary:
This implements the invite link button in the frame. We take a invite link parameter in the route. The button will redirect to this link

We also check if the invite link parameter is a valid Comm invite link by checking if it is prefixed by `https://comm.app/invite/`. If not, we default to `https://comm.app`

Depends on D13947

Test Plan:
Used the warpcast frame validator https://warpcast.com/~/developers to display the frame and confirmed button redirected to invite link provided as the parameter with the following url in the warpcast frame validator: https://<keyserver url>/https%3A%2F%2Fcomm.app%2Finvite%2Fcryptoleft

Used a different link than a Comm link and confirmed it still defaults to `https://comm.app`

Reviewers:
varun,
ashoat

Subscribers: