[CommCoreModule][native_rust_library] move constants to one place
Summary:
Part of [ENG-9656](https://linear.app/comm/issue/ENG-9656/native-changes-for-minimal-version-of-backup)

I'm not sure why it was previously defined in two places. It makes it hard to maintain, so I'm cleaning this up. I added the missing comment.

Depends on D13937

Test Plan: Tested creating backup and secondary device login

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14000