[CommCoreModule] avoid calling Backup Service to get backup secrets
Summary:
Part of [ENG-9656](https://linear.app/comm/issue/ENG-9656/native-changes-for-minimal-version-of-backup)

It was identified in https://phab.comm.dev/D13849#387643.

We don't need to call Backup Service, we have the data locally.

`retrieveBackupKeys` will be refactored later in the stack and used in Restore protocol to fetch the User Keys from the backup service.

Depends on D14001

Test Plan: Tested creating backup and secondary device login

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14002