[native] remove `BackupKeys` and use `QRAuthBackupData`
Summary:
Part of [ENG-9656](https://linear.app/comm/issue/ENG-9656/native-changes-for-minimal-version-of-backup)

`BackupKeys` can be missleading and confuse with with actual Use Keys, additionally it also contains `backupID`. This was duplicated with `QRAuthBackupData` type, which is the same type, but it's more suitable as it's going to be used only in context of Secondary Device Login protocol.

Additionally renaming `retrieveBackupKeys` -> `getQRAuthBackupData`

Depends on D14002

Test Plan: Tested creating backup and secondary device login

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14003