[keyserver/lib] Change frame image from hello world to channel image
Summary:
Currently the frame image is just the text "Hello world". As part of the first changes to the image, we're replacing the background of the image
with the farcaster channel's cover image if it exists. If it doesn't, we use the default channel image

Depends on D13971

Test Plan:
Tested in the warpcast frame validator. Confirmed that the frame image displayed was just the channel cover image

Reviewers:
varun,
ashoat

Subscribers: