[native] DisplayCommunityDirectoryPrompt
Summary:
the second time a logged in user foregrounds the app, they will see the bottom sheet asking if they want to check out the community directory. if the user is not on the home screen, they won't get the prompt

Depends on D13997

Test Plan: confirmed that prompt only displays once per installation. it doesn't display if user is not logged in or not on the home screen

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D14053