[keyserver] Handle invalid CSAT for Identity RPCs
Summary:
Address [[ https://linear.app/comm/issue/ENG-6544/add-check-for-invalid-token-to-keyserver | ENG-6544 ]]
Similiar to D13992 but for keyserver

Depends on D13992, D14028

Test Plan: Malformed token stored in the `metadata` table. When a RPC was called, the value was cleared. Subsequent service calls caused new login attempt and recreated the metadata values

Reviewers: varun, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14029