[native] Show version unsupported when opened QR screen
Summary:

The QR screen is now for dev-only, but it's theoretically possible to access it via deep link.

For now, the clients aren't able to do it because there are no such QR codes to scan, but in the future after device list launch, they might try doing it on old client versions.
We should display the unsupported version message.

Depends on D14036

Test Plan:

Changed isDev condition and opened the QR scanner screen. I was logged out and a message was displayed.

Reviewers: kamil, angelika

Subscribers: