[keyserver] Add param to getRetention to customize activity period being checked
Summary: We want to introduce new metrics that check retention based on usage in the last week rather than usage in the last day. This will be handled in the next diff, but first we want to factor out the param.

Test Plan: Ran it on my personal server and confirmed that a message appeared with the correct metrics

Reviewers: varun, will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D14040