[CommCoreModule] sing message with custom Olm Account
Summary:
[ENG-9603](https://linear.app/comm/issue/ENG-9603/make-it-possible-to-sign-message-with-olm-account-from-user-keys).

Needed to sign new device list with old primary device account.

This doesn't need to be scheduled on `cryptoThread` because it is using ephemeral `CryptoModule` but perhaps it's better to keep similar to all crypto-related operations and use dedicated thread.

Depends on D14045

Test Plan: Tested in next diff in the stack

Reviewers: bartek, tomek