[native] make it possible to test signing using Olm Account from User Keys
Summary:
[ENG-9603](https://linear.app/comm/issue/ENG-9603/make-it-possible-to-sign-message-with-olm-account-from-user-keys).

This is a test case and an example of how to use it, probably this code can be updated, exported, and used in Restore Flow later (cc. @tomek).

This is a real-life use case because we take the previous Primary Device `deviceID` from the device list history, then sign the message using Olm Account from Backup without extracting `deviceID`, and use `deviceID` from the device list to verify — so if this passes verification, both device list history and User Keys upload/download/crypto operations work together.

Depends on D14046

Test Plan: Perform what is explained in test case

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14047