[lib][web][native] Redux store to persist latest backup info
Summary:
[ENG-9700](https://linear.app/comm/issue/ENG-9700/create-redux-store-to-persist-latest-backup-info).

Native-specific store where we can put backup-related data.

I am not sure if it is better to put this in lib, even if this won't be used on the web at all and has cleaner code (e.g., reducing in master reducer, not in native specific), or leave it as it is.

Test Plan: Check if migrations works

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14049