[native] Navigate from the QR screen to the restore screen
Summary:
Introduce a wrapper over the logged out screen that can be used to choose between prompt and restore flows - this allows avoiding duplicating the code. Configure the wrapper to be used as two screens - the prompt and restore.

https://linear.app/comm/issue/ENG-9679/create-a-screen-where-user-can-choose-between-usernamepassword-and
https://linear.app/comm/issue/ENG-9958/introduce-a-stack-navigator-for-the-logged-out-screen

{F3413335}

Depends on D14067

Test Plan: A couple of scenarios - details in the video from the summary.

Reviewers: kamil, angelika, bartek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14038