[native] Navigate from the QR screen to the restore screen
Summary:
When a user clicks a link, we open the restore screen. This screen is a part of logged out modal.

https://linear.app/comm/issue/ENG-9678/introduce-a-restore-button

Depends on D13907

Test Plan: Open the app, click log in - the QR screen should show. Going back from it opens the prompt screen (with login and register buttons). Clicking a link in it opens the restore screen.

Reviewers: kamil, angelika, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14038