[native] navigate to community directory nux tip if user declines prompt or closes bottom sheet
Summary: Depends on D14053

I'm having trouble getting screen recordings to work on my computer, so I'll have to update this diff with a recording from my phone in the morning.

Test Plan: tested dismissing the bottom sheet with a swipe down gesture and declining the prompt by pressing the button. in both cases, the tip appeared after the sheet was closed. the tip appeared in the right spot and i was able to close it.

Reviewers: ashoat

Subscribers: