[native] enable backup handler for staff
Summary:
Part of [ENG-9604](https://linear.app/comm/issue/ENG-9604/update-the-backup-handler-lifecycle-for-minimal-version-of-backup) but only for staff.

Starting Rust thread responsible for uploading Backup. Right now it will be used to upload User Keys only, but in the future, it will be used for User Data and logs.

Adding an alert to log if anything went wrong.

Depends on D14048, D14050

Test Plan: Add additional log to make sure if this is started or not, test it with iOS and Android (background/foreground/background).

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14056