[native] handle invalid CSAT in Backup authenticated API
Summary:
This is implementation on [ENG-9604#comment-0ef625dd](https://linear.app/comm/issue/ENG-9604/update-the-backup-handler-lifecycle-for-minimal-version-of-backup#comment-0ef625dd).

Similar to D14024.

We need to keep this in mind to handle this also in the future, when working on User Data restore, for now we download only User Keys which are unauthenticated, and there is no code for User Data yet.

Depends on D14060

Test Plan:
1. Backup upload works
2. Mock CSAT with random string -> logged out on backup upload attempt

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14061