[web] remove password change feature
Summary:
This was discussed in [ENG-8070](https://linear.app/comm/issue/ENG-8070/regenerate-backup-on-every-password-change#comment-c64a499b)

Only the primary device should be able to change the password.

Perhaps there was some alternative to keep this for a while and disable this after the user migrated to Signed Device Lists (v2), but this is not easy and probably there is no need to complicate just to keep this screen for a ~month and then remove.

Test Plan: No option to change password in account settings screen

Reviewers: ashoat, varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D14062