[native] allow password change only on primary device
Summary:
This was discussed in [ENG-8070](https://linear.app/comm/issue/ENG-8070/regenerate-backup-on-every-password-change#comment-c64a499b)

Only the primary device should be able to change the password.

Perhaps there was some alternative to keep this for a while and disable this after the user migrated to Signed Device Lists (v2), but this is not easy and probably there is no need to complicate just to keep it for a ~month and then remove.

Hook with check was introduced in D14058

Depends on D14061, D14062

Test Plan:
1. On primary device button to change password is visible
2. On secondary device there is no button

Reviewers: ashoat, varun