[native] Extract the buttons from the logged out modal
Summary:
These buttons will be used later in the stack. In theory, we could use the primary button for most of the cases, but the SIWE button is quite unique so it's safer to keep the current styling.

Depends on D14064

Test Plan: Check if the logged out modal still looks correct.

Reviewers: kamil, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14066