[native] Migrate messageContainerStyle in useAnimatedMessageTooltipButton to Reanimate V2 API
Summary:
This is a straightforward change: use progressV2 variable, use useAnimatedStyle instead of useMemo for styles and put the interpolation in useAnimatedStyle worklet.

Depends on D14118

Test Plan: Open up text message tooltip and verify it's animated correctly. Also navigate to the thread from the tooltip and verify the transition looks right.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14119