[native] Navigate to a SIWE backup screen
Summary:
After successful login a user should be redirected to sign a backup

https://linear.app/comm/issue/ENG-8203/implement-navigation-flow-for-new-siwe-log-in

Depends on D14085

Test Plan: Connect the wallet using restore flow. The user should get navigated to the new screen. This currently fails because the authentication logic isn't connected yet and will be done as a part o https://linear.app/comm/issue/ENG-9967/connect-the-siwe-logic-with-the-ui.

Reviewers: kamil, bartek