[keyserver] replace farcaster bot reply text with frame
Summary:
This replaces the farcaster bot reply text with a frame

Depends on D14086

Test Plan:
I tested by using `ngrok` with my local keyserver and triggering a webhook event on warpcast. This caused the farcaster bot to properly reply with the frame.
Made sure that the join channel button on the frame properly redirected to a comm app invite link

Before landing I plan on testing this on my self-hosted keyserver

Reviewers:
ashoat,
varun

Subscribers: